Refactor rate limiter code and add inflight rate limiter (#2)
* Refactor rate limiter code and add inflight rate limiter

* InFlightRateLimiter: take ctx.Done into consideration and bug fix
